+++ src/ezdxf/entities/dxfgfx.py
@@ -1,8 +1,5 @@
 # Copyright (c) 2019-2020 Manfred Moitzi
 # License: MIT License
-# Created 2019-02-13
-#
-# DXFGraphic - graphical DXF entities stored in LAYOUTS and BLOCKS
 from typing import TYPE_CHECKING, Optional, Tuple, Iterable, Dict
 from ezdxf.entities import factory
 from ezdxf import options
@@ -14,8 +11,7 @@ from ezdxf.lldxf.const import (
     DXF12, DXF2000, DXF2004, DXF2007, DXF2013, DXFValueError, DXFKeyError,
     DXFTableEntryError, SUBCLASS_MARKER, DXFInvalidLineType, DXFStructureError,
 )
-from ezdxf.lldxf.tags import Tags
-from ezdxf.math import OCS, UCS, Matrix44
+from ezdxf.math import OCS, Matrix44
 from ezdxf.tools.rgb import int2rgb, rgb2int
 from ezdxf.tools import float2transparency, transparency2float
 from ezdxf.proxygraphic import load_proxy_graphic, export_proxy_graphic
@@ -23,8 +19,8 @@ from .dxfentity import DXFEntity, base_class, SubclassProcessor
 
 if TYPE_CHECKING:
     from ezdxf.eztypes import (
-        Auditor, TagWriter, BaseLayout, DXFNamespace, Vertex, Drawing
-    )
+        Auditor, TagWriter, BaseLayout, DXFNamespace, Vertex, Drawing,
+)
 
 __all__ = [
     'DXFGraphic', 'acdb_entity', 'SeqEnd', 'add_entity',
@@ -108,10 +104,6 @@ acdb_entity = DefSubclass('AcDbEntity', {
     # line) (optional), compiled by TagCompiler() to a DXFBinaryTag() objects
 })
 
-GRAPHIC_ATTRIBUTES_TO_RECOVER = {
-    attrib.code: name for name, attrib in acdb_entity.attribs.items()
-}
-
 
 class DXFGraphic(DXFEntity):
     """ Common base class for all graphic entities, a subclass of
@@ -155,19 +147,6 @@ class DXFGraphic(DXFEntity):
         if len(tags) and not r12:
             processor.log_unprocessed_tags(tags, subclass=acdb_entity.name)
         return dxf
-
-    @staticmethod
-    def recover_graphic_attributes(tags: Tags, dxf: 'DXFNamespace') -> Tags:
-        unprocessed_tags = Tags()
-        for tag in tags:
-            attrib_name = GRAPHIC_ATTRIBUTES_TO_RECOVER.get(tag.code)
-            # Don't know if the unprocessed tag is really a misplaced tag,
-            # so check if the attribute already exist!
-            if attrib_name and not dxf.hasattr(attrib_name):
-                dxf.set(attrib_name, tag.value)
-            else:
-                unprocessed_tags.append(tag)
-        return unprocessed_tags
 
     def post_new_hook(self):
         """ Post processing and integrity validation after entity creation