+++ src/ezdxf/entities/ellipse.py
@@ -83,9 +83,7 @@ class Ellipse(DXFGraphic):
             self, processor: SubclassProcessor = None) -> 'DXFNamespace':
         dxf = super().load_dxf_attribs(processor)
         if processor:
-            tags = processor.load_dxfattribs_into_namespace(dxf, acdb_ellipse)
-            if len(tags):
-                processor.log_unprocessed_tags(tags, subclass=acdb_ellipse.name)
+            processor.load_and_recover_dxfattribs(dxf, acdb_ellipse)
         return dxf
 
     def export_entity(self, tagwriter: 'TagWriter') -> None: