+++ src/ezdxf/entities/line.py
@@ -46,9 +46,7 @@ class Line(DXFGraphic):
         """
         dxf = super().load_dxf_attribs(processor)
         if processor:
-            tags = processor.load_dxfattribs_into_namespace(dxf, acdb_line)
-            if len(tags) and not processor.r12:
-                processor.log_unprocessed_tags(tags, subclass=acdb_line.name)
+            processor.load_and_recover_dxfattribs(dxf, acdb_line)
         return dxf
 
     def export_entity(self, tagwriter: 'TagWriter') -> None: