+++ src/ezdxf/entities/point.py
@@ -68,9 +68,7 @@ class Point(DXFGraphic):
         """ Loading interface. (internal API) """
         dxf = super().load_dxf_attribs(processor)
         if processor:
-            tags = processor.load_dxfattribs_into_namespace(dxf, acdb_point)
-            if len(tags) and not processor.r12:
-                processor.log_unprocessed_tags(tags, subclass=acdb_point.name)
+            processor.load_and_recover_dxfattribs(dxf, acdb_point)
         return dxf
 
     def export_entity(self, tagwriter: 'TagWriter') -> None: