+++ src/ezdxf/entities/shape.py
@@ -61,9 +61,7 @@ class Shape(DXFGraphic):
             self, processor: SubclassProcessor = None) -> 'DXFNamespace':
         dxf = super().load_dxf_attribs(processor)
         if processor:
-            tags = processor.load_dxfattribs_into_namespace(dxf, acdb_shape)
-            if len(tags) and not processor.r12:
-                processor.log_unprocessed_tags(tags, subclass=acdb_shape.name)
+            processor.load_and_recover_dxfattribs(dxf, acdb_shape)
         return dxf
 
     def export_entity(self, tagwriter: 'TagWriter') -> None: