+++ src/ezdxf/entities/text.py
@@ -145,13 +145,8 @@ class Text(DXFGraphic):
         """ Loading interface. (internal API) """
         dxf = super().load_dxf_attribs(processor)
         if processor:
-            tags = processor.load_dxfattribs_into_namespace(dxf, acdb_text, 2)
-            if len(tags) and not processor.r12:
-                processor.log_unprocessed_tags(tags, subclass=acdb_text.name)
-
-            tags = processor.load_dxfattribs_into_namespace(dxf, acdb_text2, 3)
-            if len(tags) and not processor.r12:
-                processor.log_unprocessed_tags(tags, subclass=acdb_text2.name)
+            processor.load_and_recover_dxfattribs(dxf, acdb_text, 2)
+            processor.load_and_recover_dxfattribs(dxf, acdb_text2, 3)
         return dxf
 
     def export_entity(self, tagwriter: 'TagWriter') -> None: