+++ src/ezdxf/entities/xline.py
@@ -39,10 +39,7 @@ class XLine(DXFGraphic):
             self, processor: SubclassProcessor = None) -> 'DXFNamespace':
         dxf = super().load_dxf_attribs(processor)
         if processor:
-            tags = processor.load_dxfattribs_into_namespace(dxf, acdb_xline, 2)
-            if len(tags):
-                processor.log_unprocessed_tags(
-                    tags, subclass=self.XLINE_SUBCLASS)
+            processor.load_and_recover_dxfattribs(dxf, acdb_xline, 2)
         return dxf
 
     def export_entity(self, tagwriter: 'TagWriter') -> None: