+++ tests/test_01_dxf_entities/test_112_dxfgfx.py
@@ -1,12 +1,12 @@
 # Copyright (c) 2019-2020 Manfred Moitzi
 # License: MIT License
-# created 2019-02-14
 import pytest
 import ezdxf
 
 from ezdxf.entities.dxfgfx import DXFGraphic, DXFValueError
 from ezdxf.math import Matrix44
 from ezdxf.lldxf.tags import Tags, DXFTag
+from ezdxf.entities.dxfns import recover_graphic_attributes
 
 
 @pytest.fixture
@@ -156,7 +156,7 @@ def test_recover_acdb_entity_tags():
     entity = DXFGraphic()
     tags = Tags([DXFTag(62, 1), DXFTag(8, 'Layer'), DXFTag(6, 'Linetype')])
 
-    DXFGraphic.recover_graphic_attributes(tags, entity.dxf)
+    recover_graphic_attributes(tags, entity.dxf)
     assert entity.dxf.color == 1
     assert entity.dxf.layer == 'Layer'
     assert entity.dxf.linetype == 'Linetype'
@@ -169,7 +169,7 @@ def test_recover_acdb_entity_tags_does_not_replace_existing_attribs():
     entity.dxf.linetype = 'HasLinetype'
     tags = Tags([DXFTag(62, 1), DXFTag(8, 'Layer'), DXFTag(6, 'Linetype')])
 
-    DXFGraphic.recover_graphic_attributes(tags, entity.dxf)
+    recover_graphic_attributes(tags, entity.dxf)
     assert entity.dxf.color == 7
     assert entity.dxf.layer == 'HasLayer'
     assert entity.dxf.linetype == 'HasLinetype'
@@ -179,7 +179,7 @@ def test_recover_acdb_entity_tags_ignores_unknown_tags():
     entity = DXFGraphic()
     tags = Tags([DXFTag(62, 1), DXFTag(8, 'Layer'), DXFTag(99, 'Unknown')])
 
-    unprocessed_tags = DXFGraphic.recover_graphic_attributes(tags, entity.dxf)
+    unprocessed_tags = recover_graphic_attributes(tags, entity.dxf)
     assert len(unprocessed_tags) == 1
     assert unprocessed_tags[0] == (99, 'Unknown')