+++ NEWS.md
@@ -5,6 +5,8 @@ News
 Version 0.15a0 - dev
 --------------------
 - NEW: linetype support for matplotlib drawing backend
+- NEW: recover misplaced tags of the `AcDbEntity` subclass (color, layer, 
+  linetype, ...)
 - REMOVED: deprecated `DXFEntity.transform_to_wcs()` interface, 
   use `DXFEntity.transform(ucs.matrix)`
 - REMOVED: deprecated `Hatch.edit_boundary()` context manager,