+++ NEWS.md
@@ -2,9 +2,10 @@
 News
 ====
 
-Version 0.14.1 - dev
---------------------
+Version 0.14.1b0 - 2020-09-17
+-----------------------------
 
+- Release notes: https://ezdxf.mozman.at/release-v0-14.html
 - BUGFIX: MLEADER and MLEADERSTYLE min DXF version changed to R2000
 - BUGFIX: AutoCAD ignores not existing default objects in ACDBDICTIONARYWDFLT
   and so ezdxf have to. `Auditor()` creates a place holder object as default